fixed: gcc doesn't like r35127. Thanks bobo1on1
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@35140 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/Song.cpp
@@ -79,7 +79,7 @@ void CSong::Serialize(CVariant& value)
   value["comment"] = strComment;
   value["rating"] = rating;
   value["timesplayed"] = iTimesPlayed;
-  value["karaokenumber"] = iKaraokeNumber;
+  value["karaokenumber"] = (int64_t) iKaraokeNumber;
 }
 
 void CSong::Clear()
